move reth-db-api::compress and decompress to reth-db-serialization
